refactor(metrological_streams): refactor metrological_streams.py into a package metrological_streams and include classes into streams
